backend/ocr/requirements.txt
torch>=2.1
torchvision
transformers>=4.39
accelerate
pillow
opencv-python
sentencepiece